==> notebooks/halifax-rag/1-ingestion-halifax-rag.ipynb <==
from pathlib import Path

dataset_path = Path.cwd().joinpath("datasets", "halifax_site")

import pandas as pd

training_data = pd.read_csv(
    dataset_path.joinpath("training_data.csv"), sep="|",  lineterminator='\n')

test_data = pd.read_csv(dataset_path.joinpath(
    "test_data.csv"), sep="|",  lineterminator='\n')

training_data.head()

training_data = training_data[~training_data.content.isna()
                           & ~training_data.url.isna()]

test_data = test_data[~test_data.content.isna() & ~test_data.url.isna()]

test_data.head()

test_data.shape

from haystack import Document

train_documents = [
    Document(content=str(example.content_cleaned), meta={"url": example.url}) for example in training_data.itertuples()
]

test_documents = [
    Document(content=str(example.content_cleaned), meta={"url": example.url}) for example in test_data.itertuples()
]

all_documents = train_documents + test_documents

embedding_model_id = "dunzhang/stella_en_400M_v5"

### need to come back here, and download the stuff.

from haystack import Pipeline

from haystack import Pipeline

indexing_pipeline = Pipeline()

all_documents[0]

from haystack.components.writers import DocumentWriter

from haystack_integrations.document_stores.pgvector import PgvectorDocumentStore

from src.shared.database import postgres_uri

import os

os.environ["PG_CONN_STR"] = postgres_uri

document_store = PgvectorDocumentStore(
    embedding_dimension=1024,
    vector_function="cosine_similarity",
    recreate_table=True,
    search_strategy="hnsw",
)

from haystack.document_stores.types import DuplicatePolicy

document_writer = DocumentWriter(
    document_store=document_store, policy=DuplicatePolicy.OVERWRITE)

from haystack.components.embedders import SentenceTransformersDocumentEmbedder

embedder_component = SentenceTransformersDocumentEmbedder(
    model=embedding_model_id,
    normalize_embeddings=True,
    trust_remote_code=True,

)

embedder_component.warm_up()

indexing_pipeline.add_component(
    "embedder", embedder_component)
indexing_pipeline.add_component("writer", document_writer)


indexing_pipeline.connect("embedder", "writer")

indexing_pipeline.run(data={"documents": all_documents[1:]})
