==> notebooks/halifax-rag/2.Retrieval-halifax-rag.ipynb <==
from src.rag.components.retriever import HybridRetriever

embedding_model_id = "dunzhang/stella_en_400M_v5"

from pathlib import Path

model_path = Path.cwd().joinpath("models")

embedding_model_path = model_path.joinpath(embedding_model_id ).__str__()

cross_encoder_kwargs = {
    "model_name": embedding_model_path,
    "trust_remote_code": True,
    "local_files_only": True,
    "config_kwargs": {"use_memory_efficient_attention": False, "unpad_inputs": False},
    "device": "cpu"
}


transformer_kwargs  = { "model_name_or_path": embedding_model_path,
    "trust_remote_code": True,
    "device": "cpu",
    "config_kwargs": {"use_memory_efficient_attention": False,
                   "unpad_inputs": False},
    "cache_folder": model_path}

spacy_model_id = "en_core_web_sm"

retriever = HybridRetriever(
    cross_encoder_kwargs=cross_encoder_kwargs,
    spacy_model=spacy_model_id,
    sentence_transformer_kwargs=transformer_kwargs,
    language="english"
)

query = "How does an interest only mortgage work?"


## "How can I speed up the mortgage completion process?"


#"How does an interest only mortgage work?"

# "How do you work out your LTV ratio?"

semantic_search_results = retriever.semantic_search(query=query, limit=10)

for item in semantic_search_results:
    print(item.content)
    print(10 * "-")

keywords = retriever.perform_keyword_extraction(query)

keywords

keywords_search_results = retriever.keyword_search(keywords=keywords, limit=10)

We can add a reranker but we decided to keep it simple for now.

For some reason the encoder is taking long to return.

for item in keywords_search_results:
    print(item.content)
    print(10 * "-")

# scores = retriever.cross_encoder.predict(
#    [(query, item.content) for item in keywords_search_results + semantic_search_results])

### Generator

from src.rag.components.generator import LLamaCppGeneratorComponent

prompt = "you are a helpful mortgage advisor"

model_name = "Qwen/Qwen2.5-1.5B-Instruct"

llama_cpp_generator = LLamaCppGeneratorComponent(
    api_url="http://127.0.0.1:8080",
    model_name=model_name,
    prompt=prompt
)

llama_cpp_generator._ping_api()

documents = [doc.content for doc in semantic_search_results +
             keywords_search_results]

template_input = {
    "documents": documents,
    "question": query
}

llama_cpp_generator.generate_chat_input(template_values=template_input)

response = llama_cpp_generator.run(template_values=template_input)

from pprint import pprint

pprint(response)
